Adding test project. Basic unit testing started

+++ FedExAPI/Weight.cs
@@ -3,6 +3,6 @@
     public class Weight
     {
         public EnumWeightUnit? Units { get; set; }
-        public int? Value { get; set; }
+        public double? Value { get; set; }
     }
 }
